# PERF: add ._simple_new method to masked arrays