@topper-123 2023-05-06
Ping @jbrockmendel & @rhshadrach. This is currently contained in #52788 to fix a slowdown I saw when working in that PR.